requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
PyJWT
